requirements.txt
apistar==0.5.41 
pytest-cov==2.5.1
pycodestyle==2.3.1
gunicorn==19.7.1
black==18.6b4
orator==0.9.7
psycopg2==2.7.3
wsgicors==0.7.0
python-decouple==3.1.0
PyJWT==1.6.4
passlib==1.7.1
mkdocs-material==3.0.4
